+++ src/cmd/mod.rs
@@ -1,6 +1,8 @@
 pub mod auth;
 pub mod cat;
+pub mod cat_frames;
 pub mod convert;
+pub mod ls_frames;
 pub mod play;
 pub mod session;
 pub mod upload;